// src/types/cooperative.type.ts
export type TDocument = {
  id: number | string;
  name: string;
  type: "image" | "pdf";
  url: string;
};
// ...
export type TFormattedCooperativeData = {
  id: number;
  name: string;
  registrationNumber: string;
  status: string;
  submittedDate: string;
  cooperativeCode: string;
  fiscalYear: string;
  fullNameNepali: string;
  fullNameEnglish: string;
  registeredDate: string;
  registerYear: string;
  registeredFiscalYear: string;
  classification: string;
  type: string;
  workingArea: string;
  anyBranchOffice: boolean;
  numberOfBranches: string;
  panNumber: string;
  province: string;
  district: string;
  localLevel: string;
  wardNo: string;
  houseNo: string;
  tole: string;
  fullAddress: string;
  contactMobile: string;
  contactOffice: string;
  email: string;
  website: string;
  contactPerson: string;
  contactPhone: string;
  totalShareAmount: number;
  totalShares: number;
  totalSavingAmount: number;
  totalLiabilities: number;
  loanInOperation: number;
  institutionalAssets: number;
  maleMembers: number;
  femaleMembers: number;
  totalMembers: number;
  executiveMembers: Array<{
    id: number;
    name: string;
    position: string;
    email: string;
    phone: string;
    gender: string;
    ethnicity: string;
    panNo: string;
    nid: string;
    education: string;
    province: string;
    district: string;
    localLevel: string;
    wardNo: string | number;
    address: string;
  }>;
  committeeMembers: Array<{
    id: number;
    name: string;
    position: string;
    committeeType: string;
    email: string;
    phone: string;
    gender: string;
    ethnicity: string;
    panNo: string;
    nid: string;
    education: string;
    province: string;
    district: string;
    localLevel: string;
    wardNo: string | number;
    address: string;
  }>;
  documents: TDocument[];
};